refactor: extract element and component titles into separate components
